requirements.txt
Django==4.1
Pillow==9.0.0
redis==4.3.1
environs==9.3.0
psycopg2==2.9.3
pydantic==1.10.2
django-constance==2.9.1
discord==2.1.0
requests==2.28.1
discord-webhook==1.0.0
gunicorn
sentry-sdk
